requirements.txt
# Core dependencies
openai>=1.0,<2.0
langchain>=1.2,<2.0
langchain-core>=1.2,<2.0
langchain-openai>=1.1,<2.0
langchain-community>=0.4,<0.5
langgraph
chromadb>=1.5,<2.0
sentence-transformers>=5.0,<6.0
numpy>=1.24,<3.0
pandas>=2.0,<3.0
requests>=2.31,<3.0

# Dashboard / async orchestration
streamlit>=1.33.0
plotly>=5.18.0
httpx>=0.27.0

# Common prompt-mutation tool deps (lightweight)
nltk>=3.8.1
homoglyphs>=2.0.0

# App/runtime dependencies
python-dotenv>=1.0,<2.0
fastapi>=0.117,<1.0
uvicorn>=0.42,<1.0
pydantic>=2.0,<3.0
flask>=3.0,<4.0
gunicorn>=21.2,<23.0
pypdf>=6.0,<7.0
supabase>=2.8,<3.0
